Fuses platform_analytics & platform_logs admin tools into one
