+++ main/core/Library/Installation/Updater/Updater120413.php
@@ -30,6 +30,8 @@ class Updater120413 extends Updater
         $this->deleteToolFromDB('analytics');
         $this->deleteToolFromDB('logs');
         $this->deleteToolFromDB('progression');
+        $this->deleteAdminToolFromDB('platform_analytics');
+        $this->deleteAdminToolFromDB('platform_logs');
     }
 
     private function deleteToolFromDB($toolName)
@@ -41,5 +43,16 @@ class Updater120413 extends Updater
         ';
         $this->conn->prepare($toolSql)->execute();
         $this->log("$toolName tool deleted.");
+    }
+
+    private function deleteAdminToolFromDB($toolName)
+    {
+        $this->log("Deleting $toolName admin tool...");
+        $adminToolSql = '
+            DELETE tool FROM claro_admin_tools tool
+            WHERE tool.name = "'.$toolName.'"
+        ';
+        $this->conn->prepare($adminToolSql)->execute();
+        $this->log("$toolName admin tool deleted.");
     }
 }
