+++ main/core/Resources/modules/administration/dashboard/components/logs/logs-user.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+import {UserLogList} from '#/main/core/layout/logs'
+
+const UserLogs = () =>
+  <UserLogList
+    listUrl={['apiv2_admin_tool_logs_list_users']}
+  />
+
+export {
+  UserLogs
+}
