+++ main/core/Resources/modules/administration/dashboard/store/index.js
@@ -1,0 +1,7 @@
+import {actions} from '#/main/core/administration/dashboard/store/actions'
+import {reducer} from '#/main/core/administration/dashboard/store/reducer'
+
+export {
+  actions,
+  reducer
+}
