--- main/core/Resources/modules/administration/logs/connection/store/index.js
@@ -1,5 +1,0 @@
-import {reducer} from '#/main/core/administration/logs/connection/store/reducer'
-
-export {
-  reducer
-}
