+++ main/core/Resources/translations/tools.de.json
@@ -9,8 +9,6 @@
     "logs": "",
     "my_contacts": "",
     "parameters": "",
-    "platform_analytics": "",
-    "platform_logs": "",
     "platform_parameters": "",
     "platform_plugins": "",
     "registration_to_workspace": "",
