+++ main/core/Resources/translations/tools.en.json
@@ -12,8 +12,7 @@
     "my_contacts": "My contacts",
     "organization_management": "Organization management",
     "parameters": "Settings",
-    "platform_analytics": "Dashboard",
-    "platform_logs": "Platform logs",
+    "platform_dashboard": "Dashboard",
     "platform_parameters": "Platform parameters",
     "platform_plugins": "Platform plugins",
     "registration_to_workspace": "Registration to workspaces",
