+++ main/core/Resources/translations/tools.es.json
@@ -9,8 +9,6 @@
     "logs": "Registros",
     "my_contacts": "Contactos",
     "parameters": "Parámetros",
-    "platform_analytics": "Tablero de control",
-    "platform_logs": "Registros de la plataforma",
     "platform_parameters": "Parámetros de la plataforma",
     "platform_plugins": "Plugins de la plataforma",
     "registration_to_workspace": "Inscripción a un espacio de actividades",
