+++ main/core/Resources/translations/tools.fr.json
@@ -12,8 +12,7 @@
     "my_contacts": "Contacts",
     "organization_management": "Gestion des organisations",
     "parameters": "Paramètres",
-    "platform_analytics": "Tableau de bord",
-    "platform_logs": "Suivi de la plateforme",
+    "platform_dashboard": "Tableau de bord",
     "platform_parameters": "Paramètres de la plateforme",
     "platform_plugins": "Plugins de la plateforme",
     "registration_to_workspace": "Inscription à des espaces d'activités",
