+++ main/core/Resources/translations/tools.nl.json
@@ -10,8 +10,6 @@
     "my_contacts": "Contactpersonen",
     "organization_management": "De organisaties beheren",
     "parameters": "Instellingen",
-    "platform_analytics": "Statistieken van het platform",
-    "platform_logs": "Toezicht van het platform",
     "platform_parameters": "Instellingen van het platform",
     "platform_plugins": "Plugins van het platform",
     "registration_to_workspace": "Inschrijving aan werkruimten",
