--- main/core/Resources/views/administration/analytics.html.twig
@@ -1,31 +1,0 @@
-{% extends "ClarolineCoreBundle:administration:layout.html.twig" %}
-
-{% block title %}{{ parent() ~ ' - ' ~ 'dashboard' | trans({}, 'tools') | striptags | raw }}{% endblock %}
-
-{% import 'ClarolineCoreBundle::macros.html.twig' as macros %}
-
-{% block breadcrumb %}
-    {{
-        macros.breadcrumbs([
-            {
-                'icon': 'fa fa-cog',
-                'name': 'administration'|trans({}, 'platform'),
-                'href': path('claro_admin_index')
-            },
-            {
-                'name': 'dashboard'|trans({}, 'tools'),
-                'href': ''
-            }
-        ])
-    }}
-{% endblock %}
-
-{% block content %}
-    <div class="page-container analytics-container"></div>
-{% endblock %}
-
-{% block javascripts %}
-    {{ parent() }}
-
-    <script src="{{ hotAsset('claroline-distribution-main-core-tools-admin-analytics') }}"></script>
-{% endblock %}
