--- main/core/Resources/views/administration/logs.html.twig
@@ -1,35 +1,0 @@
-{% extends "ClarolineCoreBundle:administration:layout.html.twig" %}
-
-{% block title %}{{ parent() ~ ' - ' ~ 'admin_logs' | trans({}, 'platform') | striptags | raw }}{% endblock %}
-
-{% import 'ClarolineCoreBundle::macros.html.twig' as macros %}
-
-{% block breadcrumb %}
-    {{
-        macros.breadcrumbs([
-            {
-                'icon': 'fa fa-cog',
-                'name': 'administration'|trans({}, 'platform'),
-                'href': path('claro_admin_index')
-            },
-            {
-                'name': 'admin_logs'|trans({}, 'platform'),
-                'href': ''
-            }
-        ])
-    }}
-{% endblock %}
-
-{% block content %}
-    <div
-        class="page-container logs-container"
-        data-actions="{{ actions | json_encode | raw | escape }}"
-    >
-    </div>
-{% endblock %}
-
-{% block javascripts %}
-    {{ parent() }}
-
-    <script src="{{ hotAsset('claroline-distribution-main-core-tools-admin-logs') }}"></script>
-{% endblock %}
