+++ main/core/assets.json
@@ -23,13 +23,12 @@
       "administration-toolbar": "administration/toolbar",
       "workspace-toolbar": "workspace/toolbar",
       "resource-logs": "resource/logs",
-      "tools-admin-logs": "administration/logs",
-      "tools-admin-analytics": "administration/analytics",
       "resource": "resource/bootstrap",
       "desktop-parameters-tool": "tools/desktop-parameters",
       "home": "home/bootstrap",
       "templates-management": "administration/template/bootstrap",
-      "dashboard-tool": "tools/dashboard/bootstrap"
+      "dashboard-tool": "tools/dashboard/bootstrap",
+      "dashboard-admin-tool": "administration/dashboard/bootstrap"
     }
   },
